+++ src/test/java/de/gematik/demis/validationservice/services/ProfileParserServiceTest.java
@@ -30,21 +30,16 @@ package de.gematik.demis.validationservice.services;
 import static de.gematik.demis.validationservice.util.ResourceFileConstants.EMPTY_PROFILES_PATH;
 import static de.gematik.demis.validationservice.util.ResourceFileConstants.MINIMLAL_PROFILES_PATH;
 import static de.gematik.demis.validationservice.util.ResourceFileConstants.NOT_EXISTING_PROFILES_PATH;
-import static de.gematik.demis.validationservice.util.ResourceFileConstants.TERMINOLOGY_PROFILES_PATH;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import ca.uhn.fhir.context.FhirContext;
-import de.gematik.demis.validationservice.services.terminology.remote.TerminologyServerConfigProperties;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 
 class ProfileParserServiceTest {
 
-  private final TerminologyServerConfigProperties terminologyServerConfigProperties =
-      Mockito.mock(TerminologyServerConfigProperties.class);
   private final ProfileParserService profileParserService =
-      new ProfileParserService(FhirContext.forR4Cached(), terminologyServerConfigProperties);
+      new ProfileParserService(FhirContext.forR4Cached());
 
   @Test
   void okay() {
@@ -67,26 +62,5 @@ class ProfileParserServiceTest {
   void notExistingProfilesDirectoryThrowsException() {
     assertThatThrownBy(() -> profileParserService.parseProfile(NOT_EXISTING_PROFILES_PATH))
         .isInstanceOf(IllegalArgumentException.class);
-  }
-
-  @Test
-  void featureFlagRemoteTerminologyServerEnabled() {
-    Mockito.when(terminologyServerConfigProperties.enabled()).thenReturn(true);
-    final var result = profileParserService.parseProfile(TERMINOLOGY_PROFILES_PATH);
-    assertThat(result).isNotNull();
-    final var structureDefinitions = result.structureDefinitions();
-    assertThat(structureDefinitions).isNotNull().hasSize(1);
-    assertThat(result.valueSets()).isEmpty();
-    assertThat(result.codeSystems()).isEmpty();
-    assertThat(result.withTerminologyResources()).isFalse();
-  }
-
-  @Test
-  void featureFlagRemoteTerminologyServerDisabled() {
-    Mockito.when(terminologyServerConfigProperties.enabled()).thenReturn(false);
-    final var result = profileParserService.parseProfile(TERMINOLOGY_PROFILES_PATH);
-    assertThat(result.valueSets()).isNotNull().isNotEmpty();
-    assertThat(result.codeSystems()).isNotNull().isNotEmpty();
-    assertThat(result.withTerminologyResources()).isTrue();
   }
 }